+++ packages/storybook/examples/form/SelectRow/Customize.js
@@ -10,8 +10,8 @@ function MultipleValues() {
             <SelectRow
                 multiple
                 label="Custom 'All' label"
-                asideAll="EVERYHTING SELECTED"
-                defaultValues={['opt-a', 'opt-b', 'opt-c']}>
+                asideAllLabel="EVERYHTING SELECTED"
+                defaultValue={['opt-a', 'opt-b', 'opt-c']}>
                 <Option label="Option A" value="opt-a" />
                 <Option label="Option B" value="opt-b" />
                 <Option label="Option C" value="opt-c" />
@@ -20,8 +20,8 @@ function MultipleValues() {
             <SelectRow
                 multiple
                 label="Turn off 'All' label"
-                asideAll={false}
-                defaultValues={['opt-a', 'opt-b', 'opt-c']}>
+                asideAllLabel={false}
+                defaultValue={['opt-a', 'opt-b', 'opt-c']}>
                 <Option label="Option A" value="opt-a" />
                 <Option label="Option B" value="opt-b" />
                 <Option label="Option C" value="opt-c" />
@@ -30,7 +30,7 @@ function MultipleValues() {
             <SelectRow
                 multiple
                 label="Custom 'None' label"
-                asideNone="Nothing">
+                asideNoneLabel="Nothing">
                 <Option label="Option A" value="opt-a" />
                 <Option label="Option B" value="opt-b" />
                 <Option label="Option C" value="opt-c" />
@@ -39,9 +39,9 @@ function MultipleValues() {
             <SelectRow
                 multiple
                 label="Custom separator label"
-                asideAll={false}
+                asideAllLabel={false}
                 asideSeparator=" + "
-                defaultValues={['opt-a', 'opt-b', 'opt-c']}>
+                defaultValue={['opt-a', 'opt-b', 'opt-c']}>
                 <Option label="Option A" value="opt-a" />
                 <Option label="Option B" value="opt-b" />
                 <Option label="Option C" value="opt-c" />