+++ CHANGELOG.md
@@ -5,13 +5,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Breaking
-- [Core] [Form] [ImageEditor] Peer dependency changes:
-  * Change from `@babel/runtime-corejs2` to `@babel/runtime-corejs3`.
-
 ### Changed
 - Upgrade to Babel v7.4.4 + `core-js` v3 to provide better polyfilling.
 
+### Added
+- [Core] [Form] [ImageEditor] setup `warning@4.0.3`.
+
+### Breaking
+- [Core] [Form] [ImageEditor] Peer dependency changes:
+  * Change from `@babel/runtime-corejs2` to `@babel/runtime-corejs3`.
+- [Form] `<SelectList>`:
+    - Rename prop `values` to `value`, and it receive a single value directly when is not `multiple`, and receive an array when `multiple` is true.
+    - Rename prop `defaultValues` to `defaultValue`, and it receive a single value directly when is not `multiple`, and receive an array when `multiple` is true.
+    - Rename prop `allOptionLabel` to `checkAllLabel`.
+- [Form] `<SelectRow>`:
+    - Rename prop `values` to `value`, and it receive a single value directly when is not `multiple`, and receive an array when `multiple` is true.
+    - Rename prop `defaultValues` to `defaultValue`, and it receive a single value directly when is not `multiple`, and receive an array when `multiple` is true.
+    - Rename prop `asideAll` to `asideAllLabel`.
+    - Rename prop `asideNone` to `asideNoneLabel`.
 
 ## [3.0.0]
 ### Breaking